Refactor: 커스텀 클래스를 사용한 예외처리
