+++ src/main/java/Dino/Duett/domain/member/exception/MemberException.java
@@ -1,0 +1,20 @@
+package Dino.Duett.domain.member.exception;
+
+import Dino.Duett.global.exception.CustomException;
+import Dino.Duett.global.exception.ErrorCode;
+
+public abstract class MemberException extends CustomException {
+    protected MemberException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+    public static class DuplicatePhoneNumberException extends MemberException {
+        public DuplicatePhoneNumberException() {
+            super(ErrorCode.DUPLICATE_PHONE_NUMBER);
+        }
+    }
+    public static class DuplicateKakaoIdException extends MemberException {
+        public DuplicateKakaoIdException() {
+            super(ErrorCode.DUPLICATE_KAKAO_ID);
+        }
+    }
+}
