+++ src/main/java/Dino/Duett/domain/member/service/MemberService.java
@@ -3,6 +3,7 @@ package Dino.Duett.domain.member.service;
 import Dino.Duett.domain.member.dto.MemberDto;
 import Dino.Duett.domain.member.entity.Member;
 import Dino.Duett.domain.member.enums.MemberState;
+import Dino.Duett.domain.member.exception.MemberException;
 import Dino.Duett.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -10,7 +11,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
-import static Dino.Duett.domain.member.enums.Message.DUPLICATE_PHONE_NUMBER;
+import static Dino.Duett.global.exception.ErrorCode.DUPLICATE_KAKAO_ID;
+import static Dino.Duett.global.exception.ErrorCode.DUPLICATE_PHONE_NUMBER;
 
 @Slf4j
 @Service
@@ -22,10 +24,10 @@ public class MemberService {
     public Member createMember(String phoneNumber, String kakaoId) {
         // 중복 체크
         if (memberRepository.existsByPhoneNumber(phoneNumber)) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, DUPLICATE_PHONE_NUMBER.getMessage());
+            throw new MemberException.DuplicatePhoneNumberException();
         }
         if (memberRepository.existsByKakaoId(kakaoId)) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, DUPLICATE_PHONE_NUMBER.getMessage());
+            throw new MemberException.DuplicateKakaoIdException();
         }
 
         // 멤버 생성
